feat: update logic for updating group messages

+++ connections/src/app/Store/effects/effects.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { select, Store } from '@ngrx/store';
 import { TuiDialogService } from '@taiga-ui/core';
+import { TUI_PROMPT } from '@taiga-ui/kit';
 import {
   catchError,
   EMPTY,
@@ -16,30 +17,55 @@ import {
   withLatestFrom,
 } from 'rxjs';
 import { HttpService } from 'src/app/core/services/http.service';
-import { GroupMessagesResponseBody } from 'src/app/shared/models/group-messages';
+import { GroupMessagesResponseBody } from 'src/app/shared/models/group-messages-model';
 import { ResponseGroupID } from 'src/app/shared/models/groups-model';
-import { ResponseSinceTimestamp } from 'src/app/shared/models/http-models';
-import { ProfileResponseBody } from 'src/app/shared/models/profile-models';
+import { ResponseSinceTimestamp } from 'src/app/shared/models/http-model';
+import {
+  PeopleMessagesRequestBody,
+  PeopleMessagesResponseBody,
+} from 'src/app/shared/models/people-messages-model';
+import {
+  MergedData,
+  PeopleConversationsListResponseBody,
+  PeopleItem,
+  PeopleListResponseBody,
+} from 'src/app/shared/models/people-model';
+import { ProfileResponseBody } from 'src/app/shared/models/profile-model';
+import {
+  NotificationService,
+  toastTypes,
+} from 'src/app/shared/services/notification.service';
 import {
   createGroup,
   createGroupSuccess,
   deleteGroup,
   deleteGroupSuccess,
   deleteLoginSuccess,
+  deletePeopleConversation,
+  deletePeopleConversationSuccess,
   sendGroupMessagesData,
   sendGroupMessagesDataSuccess,
+  sendPeopleMessagesData,
+  sendPeopleMessagesDataSuccess,
   setGroupListData,
   setGroupMessagesData,
   setGroupMessagesDataSuccess,
+  setPeopleConversationID,
+  setPeopleConversationIDSuccess,
+  setPeopleConversationsListData,
   setPeopleListData,
+  setPeopleMessagesData,
+  setPeopleMessagesDataSuccess,
   setProfileData,
   updateName,
 } from '../actions/actions';
 import { selectPeopleList } from '../selectors/selectors';
 import {
   getLastReceivedTimestamp,
+  getLastReceivedTimestampPeople,
+  mergePeopleAndConversationsData,
   transformUnixTimestampToReadableDate,
-} from './date-utils';
+} from './effect-utils';
 
 @Injectable()
 export class ConnectionsEffects {
@@ -116,16 +142,6 @@ export class ConnectionsEffects {
           switchMap(() => of(deleteLoginSuccess())),
           catchError(() => of({ type: 'ERROR_ACTION' })),
           finalize(() => console.log('Logout effect completed'))
-          // exhaustMap(() => {
-          //   this.dialogService
-          //     .open('You have logged out successfully!', {
-          //       label: 'Success',
-          //       size: 's',
-          //     })
-          //     .subscribe();
-          //   return of({ type: 'NO_ACTION' });
-          // }),
-          // catchError(() => of({ type: 'ERROR_ACTION' }))
         );
       })
     )
@@ -144,7 +160,6 @@ export class ConnectionsEffects {
       })
     )
   );
-
   loadPeopleListData$ = createEffect(() =>
     this.actions$.pipe(
       ofType('[People List] Set People List Data'),
@@ -152,12 +167,59 @@ export class ConnectionsEffects {
         const headers = this.createHeaders();
         return this.httpService.getPeopleList({ headers }).pipe(
           take(1),
-          map((data) => setPeopleListData({ data })),
+          mergeMap((peopleData: PeopleListResponseBody) => {
+            const mergedData: PeopleItem[] = peopleData.Items.map((item) => ({
+              ...item,
+              hasConversation: false,
+            }));
+
+            return this.httpService
+              .getPeopleConversationsList({ headers })
+              .pipe(
+                take(1),
+                map(
+                  (conversationsData: PeopleConversationsListResponseBody) => {
+                    mergedData.forEach((person) => {
+                      person.hasConversation = conversationsData.Items.some(
+                        (conversation) =>
+                          conversation.companionID.S === person.uid.S
+                      );
+                    });
+
+                    this.store.dispatch(
+                      setPeopleListData({
+                        data: { Count: mergedData.length, Items: mergedData },
+                      })
+                    );
+                    this.store.dispatch(
+                      setPeopleConversationsListData({
+                        data: conversationsData,
+                      })
+                    );
+                    return { type: '[People List] Load Data Success' };
+                  }
+                ),
+                catchError(() => EMPTY)
+              );
+          }),
           catchError(() => EMPTY)
         );
       })
     )
   );
+  // loadPeopleConversationsListData$ = createEffect(() =>
+  //   this.actions$.pipe(
+  //     ofType('[People List] Set People Conversations List Data'),
+  //     exhaustMap(() => {
+  //       const headers = this.createHeaders();
+  //       return this.httpService.getPeopleConversationsList({ headers }).pipe(
+  //         take(1),
+  //         map((data) => setPeopleConversationsListData({ data })),
+  //         catchError(() => EMPTY)
+  //       );
+  //     })
+  //   )
+  // );
 
   createGroup$ = createEffect(() =>
     this.actions$.pipe(
@@ -166,7 +228,13 @@ export class ConnectionsEffects {
         const headers = this.createHeaders();
         const params = { name: action.name };
         return this.httpService.createNewGroup({ headers }, params).pipe(
-          tap((groupID) => console.log('createGroup groupID:', groupID)),
+          tap(() => {
+            this.notificationService.initiate({
+              title: 'Success',
+              content: `Your group ${action.name} created successfully!`,
+              type: toastTypes.success,
+            });
+          }),
           map((groupID) => createGroupSuccess({ groupID, name: action.name })),
           catchError(() => EMPTY)
         );
@@ -177,22 +245,39 @@ export class ConnectionsEffects {
   deleteGroup$ = createEffect(() =>
     this.actions$.pipe(
       ofType(deleteGroup),
-      exhaustMap((action) => {
-        const headers = this.createHeaders();
-        const params = { groupID: action.groupID };
-        return this.httpService.deleteGroup({ headers }, params).pipe(
-          take(1),
-          tap(() => {
-            this.dialogService
-              .open('Your group removed successfully!', {
-                label: 'Success',
-                size: 's',
-              })
-              .subscribe();
-          }),
-          map(() => deleteGroupSuccess({ groupID: action.groupID })),
-          catchError(() => of({ type: 'ERROR_ACTION' }))
-        );
+      switchMap((action) => {
+        return this.dialogService
+          .open<boolean>(TUI_PROMPT, {
+            label: `Do you want to delete ${action.name}  group?`,
+            size: 's',
+            data: {
+              yes: 'OK',
+              no: 'No',
+            },
+          })
+          .pipe(
+            switchMap((response) => {
+              if (response) {
+                const headers = this.createHeaders();
+                const params = { groupID: action.groupID };
+
+                return this.httpService.deleteGroup({ headers }, params).pipe(
+                  take(1),
+                  tap(() => {
+                    this.notificationService.initiate({
+                      title: 'Success',
+                      content: `Your group ${action.name} removed successfully!`,
+                      type: toastTypes.success,
+                    });
+                  }),
+                  map(() => deleteGroupSuccess({ groupID: action.groupID })),
+                  catchError(() => of({ type: 'ERROR_ACTION' }))
+                );
+              } else {
+                return EMPTY;
+              }
+            })
+          );
       })
     )
   );
@@ -213,40 +298,39 @@ export class ConnectionsEffects {
           paramsID,
           paramsSince
         );
-        return (
-          this.httpService
-            // .getGroupMessages({ headers }, paramsID, paramsSince)
-            .getGroupMessages({ headers }, paramsID, paramsSince)
+        return this.httpService
+          .getGroupMessages({ headers }, paramsID, paramsSince)
 
-            .pipe(
-              take(1),
-              map((data: GroupMessagesResponseBody) => {
-                console.log('data from http in effect', data);
-                const transformedData: GroupMessagesResponseBody = {
-                  Count: data.Count,
-                  Items: data.Items.map((item) => {
-                    const authorItem = peopleList.Items.find(
-                      (person) => person.uid.S === item.authorID.S
-                    );
-                    const authorName = authorItem
-                      ? authorItem.name.S
-                      : 'Unknown Author';
+          .pipe(
+            take(1),
+            map((data: GroupMessagesResponseBody) => {
+              console.log('data from http in effect', data);
+              const transformedData: GroupMessagesResponseBody = {
+                groupID: action.groupID,
+                lastTimestampInGroup: getLastReceivedTimestamp(data),
 
-                    return {
-                      ...item,
-                      createdAt: {
-                        S: item.createdAt.S,
-                      },
-                      authorName: authorName,
-                      lastTimestamp: getLastReceivedTimestamp(data),
-                    };
-                  }),
-                };
-                return setGroupMessagesDataSuccess({ data: transformedData });
-              }),
-              catchError(() => EMPTY)
-            )
-        );
+                Count: data.Count,
+                Items: data.Items.map((item) => {
+                  const authorItem = peopleList.Items.find(
+                    (person) => person.uid.S === item.authorID.S
+                  );
+                  const authorName = authorItem
+                    ? authorItem.name.S
+                    : 'Unknown Author';
+
+                  return {
+                    ...item,
+                    createdAt: {
+                      S: item.createdAt.S,
+                    },
+                    authorName: authorName,
+                  };
+                }),
+              };
+              return setGroupMessagesDataSuccess({ data: transformedData });
+            }),
+            catchError(() => EMPTY)
+          );
       })
     )
   );
@@ -276,10 +360,146 @@ export class ConnectionsEffects {
       })
     )
   );
+  //  .........................................................................................................
+
+  createPeopleConversation$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(setPeopleConversationID),
+      exhaustMap((action) => {
+        const headers = this.createHeaders();
+        const body = { companion: action.companion };
+        console.log('effect createPeopleConversation$ ', body);
+
+        return this.httpService
+          .createPeopleConversation(body.companion, { headers })
+          .pipe(
+            tap((conversationID) =>
+              console.log(
+                'createPeopleConversation conversationID:',
+                conversationID
+              )
+            ),
+            map((conversationID) =>
+              setPeopleConversationIDSuccess({ conversationID })
+            ),
+            catchError(() => EMPTY)
+          );
+      })
+    )
+  );
+
+  loadPeopleMessagesData$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(setPeopleMessagesData),
+      withLatestFrom(this.store.pipe(select(selectPeopleList))),
+      exhaustMap(([action, peopleList]) => {
+        const headers = this.createHeaders();
+        const paramsID: PeopleMessagesRequestBody = {
+          conversationID: action.conversationID,
+        };
+        const paramsSince: ResponseSinceTimestamp = {
+          since: action.since,
+        };
+
+        console.log(
+          'params from effect loadGroupsMessagesData',
+          paramsID,
+          paramsSince
+        );
+        return this.httpService
+          .getPeopleMessages({ headers }, paramsID, paramsSince)
+
+          .pipe(
+            take(1),
+            map((data: PeopleMessagesResponseBody) => {
+              console.log('data from http in effect', data);
+              const transformedData: PeopleMessagesResponseBody = {
+                Count: data.Count,
+                Items: data.Items.map((item) => {
+                  const authorItem = peopleList.Items.find(
+                    (person) => person.uid.S === item.authorID.S
+                  );
+                  const authorName = authorItem
+                    ? authorItem.name.S
+                    : 'Unknown Author';
+
+                  return {
+                    ...item,
+                    createdAt: {
+                      S: item.createdAt.S,
+                    },
+                    authorName: authorName,
+                    lastTimestamp: getLastReceivedTimestampPeople(data),
+                  };
+                }),
+              };
+              return setPeopleMessagesDataSuccess({ data: transformedData });
+            }),
+            catchError(() => EMPTY)
+          );
+      })
+    )
+  );
+
+  sendPeopleMessageData$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(sendPeopleMessagesData),
+      exhaustMap((action) => {
+        const headers = this.createHeaders();
+        const body = {
+          conversationID: action.conversationID,
+          message: action.message,
+        };
+        console.log('effect sendPeopleMessageData', body, { headers });
+
+        return this.httpService.sendPeopleMessages(body, { headers }).pipe(
+          tap(() => console.log('after this.httpService.sendPeopleMessages')),
+          map(() =>
+            sendPeopleMessagesDataSuccess({
+              conversationID: action.conversationID,
+              authorID: action.authorID,
+              message: action.message,
+            })
+          ),
+          catchError(() => EMPTY)
+        );
+      })
+    )
+  );
+  deletePeopleConversation$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(deletePeopleConversation),
+      exhaustMap((action) => {
+        const headers = this.createHeaders();
+        const params = { conversationID: action.conversationID };
+        return this.httpService
+          .deletePeopleConversation({ headers }, params)
+          .pipe(
+            take(1),
+            tap(() => {
+              this.dialogService
+                .open('Your conversation removed successfully!', {
+                  label: 'Success',
+                  size: 's',
+                })
+                .subscribe();
+            }),
+            map(() =>
+              deletePeopleConversationSuccess({
+                conversationID: action.conversationID,
+              })
+            ),
+            catchError(() => of({ type: 'ERROR_ACTION' }))
+          );
+      })
+    )
+  );
+
   constructor(
     private actions$: Actions,
     private httpService: HttpService,
     private dialogService: TuiDialogService,
-    private store: Store
+    private store: Store,
+    private notificationService: NotificationService
   ) {}
 }
